refactor(v5): move command execution out of the god class (#878, 3/4)
Third extraction. BaseTerraformCommandHandler goes from 1990 to 1750 lines --
2302 before this series started.

src/command-executor.ts owns the four things every terraform sub-command
funnels through: the wall-clock deadline (#822, CWE-1088), the bounded
stdout/stderr capture that forces silent:true (#492, #613, #632), fmt/test's
separate small-message capture (#826), and the loc-keyed failure formatter that
folds captured diagnostics underneath the message (#821). Both capture ceilings
move with them; MAX_CAPTURED_STDOUT_BYTES stays re-exported from the handler
because Tests/ExecStdoutCaptureL0.ts imports it from there.

None of this depends on which cloud provider is configured, which is the
argument for it not living on the provider class hierarchy. All five members
turned out to be entirely stateless -- zero reads of any instance field -- so
CommandExecutor takes no constructor dependencies at all. That makes it the
cleanest of the four extractions and is why the diff is almost entirely
deletion plus call-site retargeting: 32 sites in the handler and 2 in the azure
subclass, which calls the inherited execWithTimeout for az login/az account set.

The compiler could not have caught the four things that mattered here, so they
were found by sweeping Tests/ for every textual reference before running
anything:

  - CaptureOutputProtectionClassL0 used an OVERRIDE of execWithStdoutCapture as
    its test seam. Once the method left the base class that override would have
    become a method nobody calls, and the real executor would have run in place
    of the canned output -- the seam failing open, silently. It now injects a
    StubExecutor as the collaborator instead. (A scoping pass over src/ had
    reported no overrides anywhere; that was true of the provider subclasses and
    wrong about the test tree.)
  - RunAzLoginL0 patches execWithTimeout through the prototype, twice, behind an
    `as any`. Retargeted at CommandExecutor.prototype; left on the base class
    the patch would have written a property nothing reads.
  - BaseTerraformCommandHandler became an orphaned import in that same test.
  - IExecOptions became an orphaned import in the class test -- the #920 lint
    shape for the third time in this series.

That the stub seam is genuinely wired is provable by construction rather than by
assertion: the fake ToolRunner in the class test has no execAsync, so had the
StubExecutor not replaced the collaborator, the real capture path would have
thrown a TypeError instead of passing.

Gate: 9/9 repo gates exit 0, lint 0, 786 passing / 0 failing, per-file floor
met. command-executor.js 97.93 stmts / 90.69 branch / 92.59 func / 98.85 lines
against a floor of 60. Test count is unchanged from #925 because no test was
added or removed -- only four were retargeted at the new seam, which is what a
pure move should look like.
